Added downscaling algorithm, image import, and dynamic canvas resizing (#57)
I refactored the canvas data handling to add image importing along with the new pixel-art focused downscaling algorithm. There is now a resize canvas button on the toolbar as well. Additionally I started the left hand toolbar template out with some buttons.

In summary, the pixel art importing and fixing feature which is required for the image generation feature to work is no complete.

* failing at implementing scaling algorithms plus other endeavors

* oh my gosh it works i'm gonna not mess with it now lol

* quickfix made context menu close on click

* refactored canvas data handling and got dynamic scaling option to work

* Rename downscaler.js to DownScaler.js